Prevent elasticsearch ready check race condition
The previous 'kubectl wait pod' check will fail if the operator
hasn't created the pod yet. Checking the status phase on the
elasticsearch object seems to accomplish the same thing and
doesn't have side effects depending on the operator not being
yet or not having created the elasticsearch pod yet.