Improve error handling a little bit.
Escape our `find -name` arguments so they don't get globbed.
This isn't a problem in the script based on where it normally
runs, but it was a problem when I copied and pasted the find
command elsewhere and it wasn't clear why it is was behaving
differently.

Lastly, tweak an `rm` to not hide errors (it should always
succeed).