NMS-15822: Add support for configmaps that get copied to the overlay
Since config map values can't have "/" in them, we need to
support multiple config maps (which we already do), so a
new "path" key is required for plain configmaps so we can
specify different paths the files are rooted at (this "path"
key is not required for "unzip" configmaps mentioned below).

Add an optional "unzip" flag that can be specified for a
configmap entry, and in that case, any *.zip files in the
configmap will be extracted at the root of the overlay
volume (any other files in that configmap will be ignored).

Co-authored-by: mershad-manesh <105240903+mershad-manesh@users.noreply.github.com>